+++ front/src/components/DocDebug/index.vue
@@ -515,14 +515,19 @@ export default {
         }
       }
       this.sendLoading = true
-      const targetHeaders = JSON.stringify(headers)
-      const realHeaders = Object.assign({}, headers)
-      realHeaders['target-headers'] = targetHeaders
+      let realHeaders
       let url = this.url
+      // 代理转发
       if (this.isProxy) {
-        url = this.getProxyUrl('/doc/debug/v1')
+        realHeaders = {}
+        realHeaders['target-headers'] = JSON.stringify(headers)
         realHeaders['target-url'] = this.url
+        url = this.getProxyUrl('/doc/debug/v1')
+      } else {
+        realHeaders = headers
       }
+      request.call(this, item.httpMethod, url, params, data, realHeaders, isMultipart, this.doProxyResponse, error => {
+        console.error(error)
       let req = {
         url: url,
         method: item.httpMethod,