+++ sdk/sdk-java/src/main/java/cn/torna/sdk/common/RequestForm.java
@@ -1,7 +1,6 @@
 package cn.torna.sdk.common;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class RequestForm  {
@@ -9,19 +8,9 @@ public class RequestForm  {
     private RequestMethod requestMethod = RequestMethod.POST;
     /** 请求表单内容 */
     private Map<String, Object> form;
-    /** 上传文件 */
-    private List<UploadFile> files;
 
     public RequestForm(Map<? extends String, ?> m) {
         this.form = new HashMap<>(m);
-    }
-
-    public List<UploadFile> getFiles() {
-        return files;
-    }
-
-    public void setFiles(List<UploadFile> files) {
-        this.files = files;
     }
 
     public Map<String, Object> getForm() {