+++ server/server-common/src/main/java/cn/torna/common/enums/UserInfoSourceEnum.java
@@ -1,5 +1,7 @@
 package cn.torna.common.enums;
 
+import java.util.Objects;
+
 /**
  * @author tanghc
  */
@@ -8,7 +10,18 @@ public enum UserInfoSourceEnum {
     BACKEND("backend"),
     FORM("form"),
     OAUTH("oauth"),
+    LDAP("ldap"),
     ;
+
+    public static UserInfoSourceEnum of(String source) {
+        for (UserInfoSourceEnum value : UserInfoSourceEnum.values()) {
+            if (Objects.equals(source, value.source)) {
+                return value;
+            }
+        }
+        return UserInfoSourceEnum.REGISTER;
+    }
+
     private final String source;
 
     UserInfoSourceEnum(String source) {